Split builtin function definitions from _compile.py into _builtin_funcs.py
