+++ cupyx/jit/__init__.py
@@ -4,8 +4,9 @@ from cupyx.jit._interface import threadIdx  # NOQA
 from cupyx.jit._interface import blockDim  # NOQA
 from cupyx.jit._interface import blockIdx  # NOQA
 from cupyx.jit._interface import gridDim  # NOQA
-from cupyx.jit._interface import syncthreads  # NOQA
-from cupyx.jit._interface import shared_memory  # NOQA
+
+from cupyx.jit._builtin_funcs import syncthreads  # NOQA
+from cupyx.jit._builtin_funcs import shared_memory  # NOQA
 
 
 import inspect as _inspect
